moving dialog to separate file for readability
