+++ src/pages/checkout/CheckoutPage.tsx
@@ -11,13 +11,10 @@ import {
   CardContent,
   IconButton,
   Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
 } from '@mui/material';
-import { Search, Add, Remove, Close } from '@mui/icons-material';
-import { categories, buildingCodes } from '../data/checkoutPage'; //TODO remove when SQL Is hooked up
+import { Search, Add, Remove } from '@mui/icons-material';
+import CheckoutDialog from './CheckoutDialog';
+import { categories, buildingCodes } from '../../data/checkoutPage'; //TODO remove when SQL Is hooked up
 
 type CheckoutItem = {
   id: string;
@@ -31,7 +28,6 @@ type Item = {
 }
 
 const CheckoutPage = () => {
-
   const [checkoutItems, setCheckoutItems] = useState<CheckoutItem[]>([]);
   const [openSummary, setOpenSummary] = useState(false);
   const [selectedBuildingCode, setSelectedBuildingCode] = useState('');
@@ -132,52 +128,13 @@ const CheckoutPage = () => {
         </div>
       )}
 
-      <Dialog
-        onClose={() => setOpenSummary(false)}
-        aria-labelledby="customized-dialog-title"
+      <CheckoutDialog
         open={openSummary}
-      >
-        <DialogTitle sx={{ m: 0, p: 2 }} id="customized-dialog-title">
-          <span style={{ fontSize: "1.5rem" }}>Check Out Summary</span>
-        </DialogTitle>
-        <IconButton
-          aria-label="close"
-          onClick={() => setOpenSummary(false)}
-          sx={{
-            position: 'absolute',
-            right: 8,
-            top: 8,
-            color: (theme) => theme.palette.grey[500],
-          }}
-        >
-          <Close />
-        </IconButton>
-        <DialogContent dividers style={{ width: "500px" }}>
-          {
-            checkoutItems.map((item: CheckoutItem) => (
-              <div key={item.id} style={{ display: "flex", justifyContent: "space-between", alignItems: "center", borderRadius: "10px" }}>
-                <div style={{ display: "flex", justifyContent: "space-between", width: "370px", alignItems: "center" }}>
-                  <div>
-                    <h4>{item.name}</h4>
-                  </div>
-                  {renderItemQuantityButtons(item)}
-                </div>
-                <div>
-                  <Button variant="text" onClick={() => removeItemFromCart(item.id)}>Remove</Button>
-                </div>
-              </div>
-            ))
-          }
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setOpenSummary(false)}>
-            Cancel
-          </Button>
-          <Button autoFocus >
-            Confirm
-          </Button>
-        </DialogActions>
-      </Dialog>
+        onClose={() => setOpenSummary(false)}
+        checkoutItems={checkoutItems}
+        removeItemFromCart={removeItemFromCart}
+        renderItemQuantityButtons={renderItemQuantityButtons}
+      />
     </div>
   );
 }
